feat: separate main title component
