--- src/App/pages/HomePage/HomePage.module.scss
@@ -1,16 +1,0 @@
-.slider__wrapper {
-  height: 470px;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  transition: 1.5s;
-}
-
-.slider {
-  max-width: 1140px;
-  margin: 0 auto;
-
-  &__title{
-    margin: 0;
-  }
-}
